contrib/apparmor: slight refactor
- more non-template related utilities to the apparmor.go file
- remove aaParser utility
